# [hybrid performance] optim the grad fuse for pipeline mode by sorting the grad by dtype