<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
Performance optimization

### PR changes
Others

### Describe
Sorted the grad by dtype before coalescing, can decrease the number of coalescing op.
Besides, by reducing the number of coalesce op, the number of c_allreduce_sum op can also be reduced.

_All test are based on Ernie3.0, pp=dp=mp=2_
_fp16_allreduce=True_ 
_optimize_cast=True_ 

#### Decrease in number of coalesce op
|| number before this opt | number after this opt | gain | 
|:--:|:--:|:--:|:--:|
|card 0/4| 48| 6 | -700% |
|card 1/5| 48|6| -700%|
|card 2/6| 120|12 | -900%|
|card 3/7| 120| 12| -900%|
|one dp group|336 | 36| -833%|

#### Loss diff
![Screen Shot 2021-08-23 at 10 44 10 AM](https://user-images.githubusercontent.com/25279174/130382968-fddac78e-750d-4d65-9eb6-dcc9911b41b3.png)
![Screen Shot 2021-08-23 at 10 44 24 AM](https://user-images.githubusercontent.com/25279174/130382973-f97cf702-f8cc-49a0-a5e1-79e6a9b2e06f.png)

#### Performance 
| throughput  before this opt | throughput after this opt | gain | 
|:--:|:--:|:--:|
| 39146 | 39232| + 0.2%|